requirements.txt
      
# Основные зависимости приложения
aiogram==3.20.0.post0
python-binance==1.0.28
python-dotenv==1.1.0
python-telegram-bot==22.0
python-binance==1.0.28
rich==14.0.0
tabulate==0.9.0
tradingview-ta==3.3.0

# Зависимости для тестирования (лучше вынести в requirements-dev.txt)
pytest==8.3.5

    
pytest-mock~=3.14.0